+++ src/index.js
@@ -15,9 +15,9 @@
 //                                                                            //
 ////////////////////////////////////////////////////////////////////////////////
 
-import './polyfills/index.js';
-import './d3-selectgroup.js';
-import './d3-selectdiv.js';
+import "./d3-selectdiv.js";
+import "./d3-selectgroup.js";
+import "./polyfills/index.js";
 
 ////////////////////////////////////////////////////////////////////////////////
 //                                                                            //
@@ -25,34 +25,32 @@ import './d3-selectdiv.js';
 //                                                                            //
 ////////////////////////////////////////////////////////////////////////////////
 
-
-export * from './fn.js';
-export * from './axis.js';
-export * from './aspectRatio.js';
-export * from './bounds.js';
-export * from './color.js';
-export * from './createSvgLayer.js';
-export * from './fallback.js';
-export * from './loadError.js';
-export * from './parse.js';
-export * from './behavior/index.js';
-export * from './layout/index.js';
-export * from './svgUtils/index.js';
-export * from './cascade.js';
-export * from './createHtmlLayer.js';
-export * from './format.js';
-export * from './transition.js';
-export * from './control/index.js';
-export * from './annotation/index.js';
-export * from './component/index.js';
-export * from './legend/index.js';
-export * from './responsiveProps.js';
-export * from './scale.js';
-export * from './locale.js';
-
-export * from './breakpoint.js';
-export * from './patterns.js';
-export * from './map/index.js';
-export * from './viewport/resize.js';
-
-export * from './maps/index.js';
+export * from "./annotation/index.js";
+export * from "./aspectRatio.js";
+export * from "./axis.js";
+export * from "./behavior/index.js";
+export * from "./bounds.js";
+export * from "./breakpoint.js";
+export * from "./cascade.js";
+export * from "./color.js";
+export * from "./component/index.js";
+export * from "./control/index.js";
+export * from "./createHtmlLayer.js";
+export * from "./createSvgLayer.js";
+export * from "./fallback.js";
+export * from "./fn.js";
+export * from "./format.js";
+export * from "./layout/index.js";
+export * from "./legend/index.js";
+export * from "./loadError.js";
+export * from "./locale.js";
+export * from "./map/index.js";
+export * from "./maps/index.js";
+export * from "./measure.js";
+export * from "./parse.js";
+export * from "./patterns.js";
+export * from "./responsiveProps.js";
+export * from "./scale.js";
+export * from "./svgUtils/index.js";
+export * from "./transition.js";
+export * from "./viewport/resize.js";