+++ docs/static/CHANGELOG.md
@@ -2,6 +2,13 @@
 
 - Change `formatNumber` to remove insignificant trailing zeros
 - Change `sszvis.move` to also invert pointScales
+- Added new function `measureText` to calculate the width of a string
+- Added new function `measureAxisLabel` as a preset of `measureText`
+- Added new function `measureLegendLabel` as a preset of `measureText`
+- Created new module `sszvis/measure` with utilities to measure elements
+- Added new function `legendBounds` to compute rows, columns, and widths for color legends
+- Fixed cut-off of color legend circle
+- Removed distinction between single- and multi-line chart
 
 ## 2.0 (Dec 2017)